It was developed to support attacks against Israel
The Ransomware virus was developed and improved with smart techniques so that it only works on Zionist systems - perhaps some will ask why only Zionist systems? Simply put, the nature of work with this virus is to make huge sums of money from institutions or individuals in exchange for recovering their encrypted data, but when it is blocked from all countries so that it works only in the Zionists’ systems, we will be certain that the weak of soul will not be able to use it against their brothers. Therefore, after consulting with the team leadership, it was decided to publish the virus so that it is accessible to everyone and they can use it in their attacks on the Zionists, but in the event that one of you is able to succeed in encrypting one of the Zionists’ systems The encryption will not be decrypted except by us, to ensure that weak souls do not exploit the situation to make money through it. If we want the world and its adornments, we can sell this virus with its decryption for thousands of dollars in profitable returns and bear the burdens of others, but wealth is not everything.